functions/requirements.txt
firebase_functions~=0.1.0
langchain
langchain-google-genai
pillow
google-cloud-firestore
forex-python